Display more informations for each repository
