+++ examples/async/Redux.Async.Universal/MainPage.xaml
@@ -17,11 +17,13 @@
             x:Name="SearchRepositoriesTextBox"
             KeyUp="SearchRepositoriesTextBox_KeyUp"
             Grid.Row="0"
+            IsSpellCheckEnabled="False"
             PlaceholderText="Search Repositories..."/>
 
         <ItemsControl 
             x:Name="RepositoriesItemsControl"
-            Grid.Row="1">
+            Grid.Row="1"
+            ScrollViewer.VerticalScrollBarVisibility="Auto">
             <ItemsControl.Template>
                 <ControlTemplate TargetType="ItemsControl">
                     <ScrollViewer x:Name="ScrollViewer"
@@ -43,22 +45,53 @@
             </ItemsControl.Template>
             <ItemsControl.ItemTemplate>
                 <DataTemplate>
-                    <Grid>
-                        <TextBlock 
-                            Text="{Binding FullName}"
-                            Margin="20"/>
+                    <Grid Height="120">
+                        <HyperlinkButton
+                            Content="{Binding FullName}"
+                            NavigateUri="{Binding HtmlUrl}"
+                            FontSize="16"
+                            HorizontalAlignment="Left"
+                            VerticalAlignment="Top"
+                            Margin="20,10"/>
 
                         <TextBlock 
+                            HorizontalAlignment="Left"
+                            VerticalAlignment="Bottom"
+                            Text="{Binding Description}"
+                            Margin="20,20,150,20"
+                            Foreground="#666666"
+                            TextTrimming="WordEllipsis"
+                            TextWrapping="Wrap"
+                            FontSize="13"/>
+
+                        <StackPanel
+                            Orientation="Horizontal"
                             HorizontalAlignment="Right"
+                            VerticalAlignment="Top"
                             Margin="20">
-                            <Run Text="&#xE1CF;"
-                                FontFamily="Segoe MDL2 Assets"/> 
-                            <Run Text="{Binding StargazersCount}"/>
-                            <Run Text="  "/>
-                            <Run Text="Forks"/>
-                            <Run Text="{Binding Forks}"/>
-                        </TextBlock>
-                        
+                            <TextBlock 
+                                Margin="0,0,15,0">
+                                <Run Text="&#xE1CF;"
+                                    FontFamily="Segoe MDL2 Assets"/> 
+                                <Run Text="{Binding StargazersCount}"/>
+                            </TextBlock>
+
+                            <Image 
+                                Height="16"
+                                Width="16"
+                                Source="Assets/Fork.png"/>
+
+                            <TextBlock
+                                Text="{Binding Forks}"/>
+
+                        </StackPanel>
+
+                        <TextBlock
+                            HorizontalAlignment="Right"
+                            VerticalAlignment="Bottom"
+                            Margin="20"
+                            Text="{Binding Language}"/>
+
                         <Rectangle 
                             Height="1"
                             HorizontalAlignment="Stretch"
