+++ examples/async/Redux.Async.Universal/Redux.Async.Universal.csproj
@@ -108,6 +108,7 @@
     <None Include="Redux.Async.Universal_TemporaryKey.pfx" />
   </ItemGroup>
   <ItemGroup>
+    <Content Include="Assets\Fork.png" />
     <Content Include="Properties\Default.rd.xml" />
     <Content Include="Assets\LockScreenLogo.scale-200.png" />
     <Content Include="Assets\SplashScreen.scale-200.png" />
