+++ examples/async/Redux.Async/ApplicationState.cs
@@ -1,6 +1,4 @@
-﻿using Redux.Async.Models;
-
-namespace Redux.Async
+﻿namespace Redux.Async
 {
     public class ApplicationState
     {
