--- examples/async/Redux.Async/Models/SearchRepositoriesResponse.cs
@@ -1,7 +1,0 @@
-﻿namespace Redux.Async.Models
-{
-    public class SearchRepositoriesResponse
-    {
-        public Repository[] Items { get; set; }
-    }
-}
