+++ examples/async/Redux.Async/Reducers.cs
@@ -1,6 +1,4 @@
-﻿using Redux.Async.Models;
-
-namespace Redux.Async
+﻿namespace Redux.Async
 {
     public static class Reducers
     {
